small pr to make all migrations skipable and updated last one to match existing pattern (#852)

+++ db/migrate/20210806044408_remove_unused_last_error.rb
@@ -1,7 +1,7 @@
 class RemoveUnusedLastError < ActiveRecord::Migration[5.1]
   def change
-    remove_column :bulkrax_entries, :last_error
-    remove_column :bulkrax_exporters, :last_error
-    remove_column :bulkrax_importers, :last_error
+    remove_column :bulkrax_entries, :last_error if column_exists?(:bulkrax_entries, :last_error)
+    remove_column :bulkrax_exporters, :last_error if column_exists?(:bulkrax_exporters, :last_error)
+    remove_column :bulkrax_importers, :last_error if column_exists?(:bulkrax_importers, :last_error)
   end
 end
